+++ alitheia/core/pom.xml
@@ -176,8 +176,7 @@
     	<version>1.9.5</version>
     	<scope>test</scope>
     </dependency>
-  </dependencies>
-    
+</dependencies>
   <profiles>
     <profile>
       <id>tools.jar</id>
