+++ alitheia/core/src/main/java/eu/sqooss/impl/service/metricactivator/MetricActivatorImpl.java
@@ -33,15 +33,26 @@
 
 package eu.sqooss.impl.service.metricactivator;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Random;
+import java.util.Set;
+import java.util.SortedSet;
+import java.util.TreeSet;
 import java.util.concurrent.atomic.AtomicLong;
 
-import eu.sqooss.service.abstractmetric.InvocationOrder;
 import org.osgi.framework.BundleContext;
 
 import eu.sqooss.core.AlitheiaCore;
-import eu.sqooss.service.abstractmetric.AbstractMetric;
+import eu.sqooss.service.abstractmetric.DefaultMetric;
 import eu.sqooss.service.abstractmetric.AlitheiaPlugin;
+import eu.sqooss.service.abstractmetric.InvocationOrder;
 import eu.sqooss.service.abstractmetric.SchedulerHints;
 import eu.sqooss.service.cluster.ClusterNodeActionException;
 import eu.sqooss.service.cluster.ClusterNodeService;
@@ -68,7 +79,9 @@ import eu.sqooss.service.pa.PluginInfo;
 import eu.sqooss.service.scheduler.Job;
 import eu.sqooss.service.scheduler.Scheduler;
 import eu.sqooss.service.scheduler.SchedulerException;
-import eu.sqooss.service.util.GraphTS;
+import eu.sqooss.service.util.Graph;
+import eu.sqooss.service.util.GraphSorter;
+import eu.sqooss.service.util.TopologicalGraphSorter;
 
 public class MetricActivatorImpl  implements MetricActivator {
 
@@ -91,7 +104,7 @@ public class MetricActivatorImpl  implements MetricActivator {
     @Override
 	public <T extends DAObject> void runMetric(T resource, AlitheiaPlugin ap) {
     	Class<? extends DAObject> activator = resource.getClass();
-    	Job j = new MetricActivatorJob((AbstractMetric)ap, resource.getId(), logger, 
+    	Job j = new MetricActivatorJob((DefaultMetric)ap, resource.getId(), logger, 
     			metricTypesToActivators.get(activator),
     			priority.incrementAndGet(),
     			fastSync);
@@ -118,7 +131,7 @@ public class MetricActivatorImpl  implements MetricActivator {
         
         /* Fire up plug-ins */
         for (PluginInfo pi : plugins) {
-           AbstractMetric m = (AbstractMetric) bc.getService(pi.getServiceRef());
+           DefaultMetric m = (DefaultMetric) bc.getService(pi.getServiceRef());
            try {
                sched.enqueue(new MetricSchedulerJob(m, sp));
            } catch (SchedulerException e) {
@@ -224,7 +237,8 @@ public class MetricActivatorImpl  implements MetricActivator {
     	Map<AlitheiaPlugin, Integer> idx = new HashMap<AlitheiaPlugin, Integer>();
     	Map<Integer, AlitheiaPlugin> invidx = new HashMap<Integer, AlitheiaPlugin>();
     	
-    	GraphTS<AlitheiaPlugin> graph = new GraphTS<AlitheiaPlugin>(unordered.size());
+    	Graph<AlitheiaPlugin> graph = new Graph<AlitheiaPlugin>(unordered.size());
+    	GraphSorter<AlitheiaPlugin> sorter = new TopologicalGraphSorter<AlitheiaPlugin>(graph);
     	
     	//Build the adjacency matrix
     	for (AlitheiaPlugin p : unordered) {
@@ -252,7 +266,7 @@ public class MetricActivatorImpl  implements MetricActivator {
     	    }
     	}
     	
-    	List<AlitheiaPlugin> sorted = graph.topo();
+    	List<AlitheiaPlugin> sorted = sorter.sort();
     	
     	logger.debug("Calculated metric order:");
     	for (AlitheiaPlugin p : sorted) {
@@ -314,8 +328,8 @@ public class MetricActivatorImpl  implements MetricActivator {
             	}
             }
             
-            AbstractMetric metric = 
-                (AbstractMetric) bc.getService(mi.getServiceRef());
+            DefaultMetric metric = 
+                (DefaultMetric) bc.getService(mi.getServiceRef());
             HashSet<Job> jobs = new HashSet<Job>();
             
             /*Check what is the default activation ordering as suggested by the metric*/
