+++ alitheia/core/src/main/java/eu/sqooss/service/abstractmetric/AlitheiaPlugin.java
@@ -96,215 +96,6 @@ import eu.sqooss.service.scheduler.Job;
  *  </ul>
  *
  */
-public interface AlitheiaPlugin {
+public interface AlitheiaPlugin extends MetricMetaData, MetricMeasurementEvaluation, MetricLifeCycle, MetricConfiguration{
 
-    /**
-     * Get the metric version. Free form text.
-     *
-     * @return The metric's version.
-     */
-    String getVersion();
-
-    /**
-     * Get information about the metric author
-     *
-     * @return The metric's author.
-     */
-    String getAuthor();
-
-    /**
-     * Get the date this version of the metric has been installed
-     *
-     * @return The metric's installation date.
-     */
-    Date getDateInstalled();
-
-    /**
-     * Get the metric name
-     *
-     * @return The metric's name.
-     */
-    String getName();
-
-    /**
-     * Get a free text description of what this metric calculates
-     *
-     * @return The metric's description.
-     */
-    String getDescription();
-
-    /**
-     * Get the metric result, without triggering a metric recalculation if
-     * the result is not present.
-     * If the result was not calculated yet, the result set is empty. If you
-     * want to trigger the calculation to get a result, use getResult() instead.
-     *
-     * @param o DAO whose type specifies the specialized sub-interface to use
-     *          and whose value determines which result to get.
-     * @return l A list of metrics
-     * @return value of the measurement or null if there is no such measurement.
-     * @throws MetricMismatchException if the DAO type is one not supported by
-     *          this metric.
-     */
-    List<Result> getResultIfAlreadyCalculated(DAObject o, List<Metric> l)
-    	throws MetricMismatchException;
-
-    /**
-     * Get a metric result. 
-     * If the result was not calculated yet, the plugin's run method is called,
-     * and the request waits until the run method returns.
-     * If you don't want this behavior, use getResultIfAlreadyCalculated()
-     * instead.
-     *
-     * @param o DAO whose type specifies the specialized sub-interface to use
-     *          and whose value determines which result to get.
-     * @return l A list of metrics
-     * @return value of the measurement or null if there is no such measurement.
-     * @throws MetricMismatchException if the DAO type is one not supported by
-     *          this metric.
-     * @throws AlreadyProcessingException to signify that the provided DAO is 
-     * currently locked for processing by another thread.  
-     * @throws Exception All exceptions initiated by the errors in code 
-     * included in implemenations of those classes.           
-     */
-    List<Result> getResult(DAObject o, List<Metric> l)
-        throws MetricMismatchException, AlreadyProcessingException, Exception;
-
-    /**
-     * Get the description objects for all metrics supported by this plug-in
-     * as found in the database.
-     *
-     * @return the list of metric descriptors, or null if none
-     */
-    List<Metric> getAllSupportedMetrics();
-
-    /**
-     * Get all metrics that are bound to the provided activation type. 
-     *
-     * @return the list of metric DAOs for the provided activation type, empty
-     * if the plug-in does not support the provided activation type
-     */
-    List<Metric> getSupportedMetrics(Class<? extends DAObject> activationType);
-
-    /**
-     * This method performs a measurement for
-     * the given DAO, if possible. The DAO might be any one of the types
-     * that make sense for measurements -- ProjectVersion, projectFile,
-     * some others. If a DAO of a type that the metric doesn't support
-     * is passed in, throws a MetricMismatchException.
-     *
-     * The calculation of measurements may be a computationally expensive
-     * task, so metrics should start jobs (by themselves) to handle that.
-     * The subclass AbstractMetric handles job creation automatically for
-     * metrics that have simple requirements (a single job for doing the
-     * calculation).
-     *
-     * Note that even if you use (parallel running) jobs in your jobs, the
-     * metric's run method needs to block until the result is calculated.
-     *
-     * @param o The DAO that gets passed to the plug-in in order to run it
-     * @throws MetricMismatchException if the DAO is of an unsupported type.
-     * @throws AlreadyProcessingException to signify that the provided DAO is 
-     * currently locked for processing by another thread. 
-     * @throws Exception Any other exception initiated from the plugin code
-     */
-    void run(DAObject o) throws MetricMismatchException, 
-        AlreadyProcessingException, Exception;
-
-    /**
-     * After installing a new version of the metric, try to
-     * update the results. The metric may opt to partially
-     * or fully update its results tables or files.
-     *
-     * @return True, if the update succeeded, false otherwise
-     */
-    boolean update();
-
-    /**
-     * Perform maintenance operations when installing a new
-     * version of the metric
-     *
-     * @return True if installation succeeded, false otherwise
-     */
-    boolean install();
-
-    /**
-     * Free the used resources and clean up on metric removal
-     *
-     * @return True, if the removal succeeded, false otherwise
-     */
-    boolean remove();
-
-    /**
-     * Clean results on project removal
-     * 
-     * @param sp The DAO to be used as reference when cleaning up results.
-     * @return True, if the cleanup succeeded, false otherwise
-     */
-    boolean cleanup(DAObject sp);
-
-    /**
-     * Return a string that is unique for this plugin, used for indexing this
-     * plugin to the system database
-     *
-     * @return A unique string, max length 255 characters
-     */
-    String getUniqueKey();
-
-    /**
-     * Get the types supported by this plug-in for data processing and result
-     * retrieval. An activation type is DAO subclass which is passed as argument
-     * to the {@link AlitheiaPlugin.run()} and
-     * {@link AlitheiaPlugin.getResult()}} methods to trigger metric
-     * calculation and result retrieval.
-     *
-     * @return A set of DAObject subclasses
-     */
-    Set<Class<? extends DAObject>> getActivationTypes();
-
-    /**
-     * Get the activation type that corresponds to the activation type which 
-     * the metric result is stored. 
-     * 
-     * @param m - The metric for which to search for an activation type
-     * @return A list of subclasses of DAObject (a.k.a activation types).
-     */
-    List<Class<? extends DAObject>> getMetricActivationTypes (Metric m);
-
-    /**
-     * Retrieves the list of configuration properties for this plug-in.
-     * <br/>
-     * Metric plug-ins can use the <code>AbstractMetric</code>'s
-     * <code>addConfigEntry</code> and <code>removeConfigEntry</code> methods
-     * to manage their own configuration schema.
-     *
-     * @return The set of the existing configuration properties for
-     *   this plug-in. This may be an empty list if no configuration is
-     *   needed or if the plug-in is not active.
-     */
-    Set<PluginConfiguration> getConfigurationSchema();
-
-    /**
-     * Metric mnemonics for the metrics required to be present for this 
-     * plugin to operate. 
-     * 
-     * @return A, possibly empty, set of metric mnemonics. 
-     */
-    Set<String> getDependencies();
-
-    /**
-     * Get a list of object ids for the database entities to run the metric
-     * on, ordered by activation type. This method essentially allows the plugin
-     * to specify a custom processing order for metadata entities to be processed
-     * by metrics. The default execution order is specified 
-     * 
-     */
-    Map<MetricType.Type, SortedSet<Long>> getObjectIdsToSync(StoredProject sp, Metric m) 
-    	throws MetricActivationException;
-
-    /**
-     * Set a reference to the scheduler job that executes the metric. The
-     * job reference is only set when the metric's execute method is called.
-     */
-    void setJob(Job j);
 }
