+++ alitheia/core/src/main/java/eu/sqooss/service/util/GraphSorter.java
@@ -1,0 +1,21 @@
+package eu.sqooss.service.util;
+
+import java.util.List;
+
+/**
+ * Abstract class for sorting (and flattening) a Graph
+ * @param <T>
+ */
+public abstract class GraphSorter<T> {
+	protected Graph<T> graph;
+	
+	public GraphSorter(Graph<T> graph){
+		this.graph = graph;
+	}
+	
+	/**
+	 * Sort and flatten the Graph into a List
+	 */
+	public abstract List<T> sort();
+	
+}
