+++ alitheia/core/src/test/java/eu/sqooss/core/MockServiceExtended.java
@@ -1,0 +1,6 @@
+package eu.sqooss.core;
+
+
+public class MockServiceExtended extends MockService {
+
+}
