+++ alitheia/core/src/test/java/eu/sqooss/test/service/scheduler/Testester.java
@@ -1,0 +1,13 @@
+package eu.sqooss.test.service.scheduler;
+
+import org.junit.Test;
+import static org.junit.Assert.*;
+
+public class Testester {
+	@Test
+	public void testTester(){
+		assertTrue(true);
+		System.out.println("yesyesyes");
+		assertTrue(false);
+	}
+}
