+++ metrics/contrib/maven-eclipse.xml
@@ -1,0 +1,8 @@
+<project default="copy-resources">
+  <target name="init"/>
+  <target name="copy-resources" depends="init">
+    <copy todir="target/classes/eu/sqooss/service/db" filtering="false">
+      <fileset dir="src/main/java/eu/sqooss/metrics/contrib/db" includes="**/*.xml" excludes="**/*.java"/>
+    </copy>
+  </target>
+</project>
